src/requirements.txt
flask>=2.0.0
stripe>=5.0.0
mcrcon>=0.7.0
requests>=2.28.0
python-dotenv>=1.0.0
